Added trans and iter for merge and quick

+++ BaseSort.cpp
@@ -1,7 +1,7 @@
 #include "BaseSort.h"
 //array = (double *)malloc(size * sizeof(double));
 
-int BaseSort::getSortTime(){
+long int BaseSort::getSortTime(){
     return sort_time;
 }
