+++ BaseSort.h
@@ -6,13 +6,13 @@ class BaseSort
 protected:
     int iteration_count;
     int transposition_count;
-    int sort_time; 
+    long int sort_time; 
     
     //double * getSortArray();
 public:
     int getIterationCount();
     int getTranspositionCount();
-    int getSortTime();
+    long int getSortTime();
     virtual double* AlgSort(double array[], int num);
 };
